update!
added in a movie maker function for when you want to create a movie after the fact - still need to figure out a way to get the transform into the function elegantly. ...

also started playing around with arbitrate weights, seeing what there is to see.

oh yeah, the simulation is actually stable now!!!